requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Werkzeug==2.3.7
pymongo==4.6.1
web3==6.15.1
qrcode==7.4.2
Pillow==10.2.0
